Modularize demo classes that are useful for making demos
