+++ fetch_gazebo_demo/scripts/demo.py
@@ -28,219 +28,9 @@
 
 # Author: Michael Ferguson
 
-import copy
-import actionlib
 import rospy
 
-from math import sin, cos
-from moveit_python import (MoveGroupInterface,
-                           PlanningSceneInterface,
-                           PickPlaceInterface)
-from moveit_python.geometry import rotate_pose_msg_by_euler_angles
-
-from control_msgs.msg import FollowJointTrajectoryAction, FollowJointTrajectoryGoal
-from control_msgs.msg import PointHeadAction, PointHeadGoal
-from grasping_msgs.msg import FindGraspableObjectsAction, FindGraspableObjectsGoal
-from geometry_msgs.msg import PoseStamped
-from move_base_msgs.msg import MoveBaseAction, MoveBaseGoal
-from moveit_msgs.msg import PlaceLocation, MoveItErrorCodes
-from trajectory_msgs.msg import JointTrajectory, JointTrajectoryPoint
-
-# Move base using navigation stack
-class MoveBaseClient(object):
-
-    def __init__(self):
-        self.client = actionlib.SimpleActionClient("move_base", MoveBaseAction)
-        rospy.loginfo("Waiting for move_base...")
-        self.client.wait_for_server()
-
-    def goto(self, x, y, theta, frame="map"):
-        move_goal = MoveBaseGoal()
-        move_goal.target_pose.pose.position.x = x
-        move_goal.target_pose.pose.position.y = y
-        move_goal.target_pose.pose.orientation.z = sin(theta/2.0)
-        move_goal.target_pose.pose.orientation.w = cos(theta/2.0)
-        move_goal.target_pose.header.frame_id = frame
-        move_goal.target_pose.header.stamp = rospy.Time.now()
-
-        # TODO wait for things to work
-        self.client.send_goal(move_goal)
-        self.client.wait_for_result()
-
-# Send a trajectory to controller
-class FollowTrajectoryClient(object):
-
-    def __init__(self, name, joint_names):
-        self.client = actionlib.SimpleActionClient("%s/follow_joint_trajectory" % name,
-                                                   FollowJointTrajectoryAction)
-        rospy.loginfo("Waiting for %s..." % name)
-        self.client.wait_for_server()
-        self.joint_names = joint_names
-
-    def move_to(self, positions, duration=5.0):
-        if len(self.joint_names) != len(positions):
-            print("Invalid trajectory position")
-            return False
-        trajectory = JointTrajectory()
-        trajectory.joint_names = self.joint_names
-        trajectory.points.append(JointTrajectoryPoint())
-        trajectory.points[0].positions = positions
-        trajectory.points[0].velocities = [0.0 for _ in positions]
-        trajectory.points[0].accelerations = [0.0 for _ in positions]
-        trajectory.points[0].time_from_start = rospy.Duration(duration)
-        follow_goal = FollowJointTrajectoryGoal()
-        follow_goal.trajectory = trajectory
-
-        self.client.send_goal(follow_goal)
-        self.client.wait_for_result()
-
-# Point the head using controller
-class PointHeadClient(object):
-
-    def __init__(self):
-        self.client = actionlib.SimpleActionClient("head_controller/point_head", PointHeadAction)
-        rospy.loginfo("Waiting for head_controller...")
-        self.client.wait_for_server()
-
-    def look_at(self, x, y, z, frame, duration=1.0):
-        goal = PointHeadGoal()
-        goal.target.header.stamp = rospy.Time.now()
-        goal.target.header.frame_id = frame
-        goal.target.point.x = x
-        goal.target.point.y = y
-        goal.target.point.z = z
-        goal.min_duration = rospy.Duration(duration)
-        self.client.send_goal(goal)
-        self.client.wait_for_result()
-
-# Tools for grasping
-class GraspingClient(object):
-
-    def __init__(self):
-        self.scene = PlanningSceneInterface("base_link")
-        self.pickplace = PickPlaceInterface("arm", "gripper", verbose=True)
-        self.move_group = MoveGroupInterface("arm", "base_link")
-
-        find_topic = "basic_grasping_perception/find_objects"
-        rospy.loginfo("Waiting for %s..." % find_topic)
-        self.find_client = actionlib.SimpleActionClient(find_topic, FindGraspableObjectsAction)
-        self.find_client.wait_for_server()
-
-    def updateScene(self):
-        # find objects
-        goal = FindGraspableObjectsGoal()
-        goal.plan_grasps = True
-        self.find_client.send_goal(goal)
-        self.find_client.wait_for_result(rospy.Duration(5.0))
-        find_result = self.find_client.get_result()
-
-        # remove previous objects
-        for name in self.scene.getKnownCollisionObjects():
-            self.scene.removeCollisionObject(name, False)
-        for name in self.scene.getKnownAttachedObjects():
-            self.scene.removeAttachedObject(name, False)
-        self.scene.waitForSync()
-
-        # insert objects to scene
-        idx = -1
-        for obj in find_result.objects:
-            idx += 1
-            obj.object.name = "object%d"%idx
-            self.scene.addSolidPrimitive(obj.object.name,
-                                         obj.object.primitives[0],
-                                         obj.object.primitive_poses[0],
-                                         wait = False)
-
-        for obj in find_result.support_surfaces:
-            # extend surface to floor, and make wider since we have narrow field of view
-            height = obj.primitive_poses[0].position.z
-            obj.primitives[0].dimensions = [obj.primitives[0].dimensions[0],
-                                            1.5,  # wider
-                                            obj.primitives[0].dimensions[2] + height]
-            obj.primitive_poses[0].position.z += -height/2.0
-
-            # add to scene
-            self.scene.addSolidPrimitive(obj.name,
-                                         obj.primitives[0],
-                                         obj.primitive_poses[0],
-                                         wait = False)
-
-        self.scene.waitForSync()
-
-        # store for grasping
-        self.objects = find_result.objects
-        self.surfaces = find_result.support_surfaces
-
-    def getGraspableCube(self):
-        graspable = None
-        for obj in self.objects:
-            # need grasps
-            if len(obj.grasps) < 1:
-                continue
-            # check size
-            if obj.object.primitives[0].dimensions[0] < 0.05 or \
-               obj.object.primitives[0].dimensions[0] > 0.07 or \
-               obj.object.primitives[0].dimensions[0] < 0.05 or \
-               obj.object.primitives[0].dimensions[0] > 0.07 or \
-               obj.object.primitives[0].dimensions[0] < 0.05 or \
-               obj.object.primitives[0].dimensions[0] > 0.07:
-                continue
-            # has to be on table
-            if obj.object.primitive_poses[0].position.z < 0.5:
-                continue
-            return obj.object, obj.grasps
-        # nothing detected
-        return None, None
-
-    def getSupportSurface(self, name):
-        for surface in self.support_surfaces:
-            if surface.name == name:
-                return surface
-        return None
-
-    def getPlaceLocation(self):
-        pass
-
-    def pick(self, block, grasps):
-        success, pick_result = self.pickplace.pick_with_retry(block.name,
-                                                              grasps,
-                                                              support_name=block.support_surface,
-                                                              scene=self.scene)
-        self.pick_result = pick_result
-        return success
-
-    def place(self, block, pose_stamped):
-        places = list()
-        l = PlaceLocation()
-        l.place_pose.pose = pose_stamped.pose
-        l.place_pose.header.frame_id = pose_stamped.header.frame_id
-
-        # copy the posture, approach and retreat from the grasp used
-        l.post_place_posture = self.pick_result.grasp.pre_grasp_posture
-        l.pre_place_approach = self.pick_result.grasp.pre_grasp_approach
-        l.post_place_retreat = self.pick_result.grasp.post_grasp_retreat
-        places.append(copy.deepcopy(l))
-        # create another several places, rotate each by 90 degrees in yaw direction
-        l.place_pose.pose = rotate_pose_msg_by_euler_angles(l.place_pose.pose, 0, 0, 1.57)
-        places.append(copy.deepcopy(l))
-        l.place_pose.pose = rotate_pose_msg_by_euler_angles(l.place_pose.pose, 0, 0, 1.57)
-        places.append(copy.deepcopy(l))
-        l.place_pose.pose = rotate_pose_msg_by_euler_angles(l.place_pose.pose, 0, 0, 1.57)
-        places.append(copy.deepcopy(l))
-
-        success, place_result = self.pickplace.place_with_retry(block.name,
-                                                                places,
-                                                                scene=self.scene)
-        return success
-
-    def tuck(self):
-        joints = ["shoulder_pan_joint", "shoulder_lift_joint", "upperarm_roll_joint",
-                  "elbow_flex_joint", "forearm_roll_joint", "wrist_flex_joint", "wrist_roll_joint"]
-        pose = [1.32, 1.40, -0.2, 1.72, 0.0, 1.66, 0.0]
-        while not rospy.is_shutdown():
-            result = self.move_group.moveToJointPosition(joints, pose, 0.02)
-            if result.error_code.val == MoveItErrorCodes.SUCCESS:
-                return
+from fetch_gazebo_demo.democlients import *
 
 if __name__ == "__main__":
     # Create a node
